#4092 improving code and cleanup
https://github.com/metasfresh/metasfresh/issues/4092
